+++ packages/embed/src/index.ts
@@ -1,3 +1,5 @@
+export * from './factories/create-popover'
 export * from './factories/create-popup'
 export * from './factories/create-slider'
 export * from './factories/create-widget'
+export * from './factories/create-sidetab'