+++ packages/embed/src/initializers/index.ts
@@ -1,3 +1,5 @@
+export * from './initialize-popovers'
 export * from './initialize-popups'
 export * from './initialize-sliders'
 export * from './initialize-widgets'
+export * from './initialize-sidetabs'